pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mmwavelab-dca1000"
version = "0.1.0"
description = "Automation and compatibility tools for TI DCA1000 + xWR mmWave raw ADC capture."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "KJ" }]
keywords = ["mmwave", "dca1000", "iwr1843", "radar", "lvds", "adc"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering",
]
dependencies = ["pyserial>=3.5"]

[project.optional-dependencies]
dev = ["pytest>=8"]

[project.scripts]
mmwl-dca1000 = "mmwavelab_dca1000.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
